server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain==0.2.17
langchain-community
langchain-chroma
langchain-core
langchain-groq
langchain-text-splitters
sentence-transformers
pypdf
chromadb
python-dotenv
pysqlite3-binary
posthog<3.0.0